Adding function to coalesce matching columns after join.
For some units, a certain column may be filled in in the emissions table and not in the facilities table (or vice vera). This can cause issues with joins and/or potentially lead to unnecessary missing data. I add a function that compares all columns from table x and y after join and if x is missing, keeps the value from y. This fixes issues with joins and keeps all available values.