Adding old unit file code to track refactoring.
This is the inital commit of code that was written for the creation of the unit file in the fall of 2023. This code is largely a one-to-one translation of the queries in Access (including naming schemes), and thus needs to be refactored to be more efficient and sensible in R. I am replacing the first few commits of the new unit file code with the old code, and committing updating chunks to track the changes during the refactoring process.