Streamlining process to gap fill heat input for "ozone season reporters."
This part is drastically different in an effort to reduce the number of steps. The key difference is that all plants with any ozone season reporting units are treated the same. That is, instead of creating two separate groups of plants (all OS reporters vs. OS and Annual reporters), they are all kept within the same dataframe, and the distribution process is the same. Then, a condition is added for when the annual heat input is updated, so that only the ozone-season-reporting units are updated. The actual distribution process is different, too, where all non-ozone season heat input is determined by distributing the non-ozone heat at the plant/prime mover level from the 923 generation and fuel file. This process, while far simpler, may be incorrect for the ozone and annual reporters. We will need to revisit and confrim that the new streamlined approach is sensible.