Updating/simplyifing process to estimate annual NOx emissions for OS reporters.
Major differences include: 1) Only including Annual reporters for the purposes of of calculating distributional proportions, afterwhich, these units are removed. And, 2) simplifying/removing aggregation steps. Some times values would be summed to lower levels only to be aggregated up. I'm not sure if the lower-level aggregations are necessary, so removing. 3) When joining the Emission Factors table in, we get multiple matches for some prime_mover, botfirty, primary_fuel_type combos. This is because some of the EFs are in different units. I added a condition to filter to only include EFs in the "PhysicalUnits" Nox unit, which fixes the issue of having extra rows added from join.