Adding function definition scripts with Roxygen comments for documentation.
Because these scripts don't live in project environment, roxygen can't be implemented normally. Instead, we use the docstring package to render function documentation. Can be accessed with ?function or docstring::docstring(function). Note that docstring evaluations the roxygen-style commens within the function definition, and not in the preamble, like normal roxygen comments